@@ -24,6 +24,7 @@
         "jss-preset-default": "^10.0.0",
         "lodash-es": "^4.17.15",
         "module-alias": "^2.2.2",
+        "rxjs": "^7.1.0",
         "tsconfig-paths": "^3.9.0"
     },
     "devDependencies": {
@@ -32,9 +33,9 @@
         "html-webpack-plugin": "4.5.0",
         "nodemon": "^2.0.7",
         "recast": "^0.19.0",
-        "ts-loader": "8.0.7",
+        "ts-loader": "^9.2.2",
         "ts-node": "^9.1.1",
-        "typescript": "4.0.5",
+        "typescript": "^4.2.4",
         "webpack": "^5.3.2",
         "webpack-cli": "^4.1.0",
         "webpack-dev-server": "3.11.0"
