@@ -18,17 +18,17 @@
   },
   "homepage": "https://github.com/Ni55aN/easyhard.js#readme",
   "devDependencies": {
-    "@rollup/plugin-typescript": "6.1.0",
-    "@testing-library/jest-dom": "^5.11.6",
-    "@types/jest": "^26.0.15",
-    "@types/ws": "^7.4.1",
-    "@typescript-eslint/eslint-plugin": "4.6.0",
-    "@typescript-eslint/parser": "4.6.0",
+    "@rollup/plugin-typescript": "^8.2.1",
+    "@testing-library/jest-dom": "^5.12.0",
+    "@types/jest": "^26.0.23",
+    "@types/ws": "^7.4.4",
+    "@typescript-eslint/eslint-plugin": "^4.24.0",
+    "@typescript-eslint/parser": "^4.24.0",
     "commitizen": "4.2.2",
     "core-js": "^3.7.0",
     "cz-conventional-changelog": "3.3.0",
-    "eslint": "7.12.1",
-    "eslint-loader": "4.0.2",
+    "eslint": "^7.27.0",
+    "eslint-loader": "^4.0.2",
     "fast-check": "^2.7.0",
     "fromentries": "^1.2.1",
     "fs-extra": "^9.0.1",
@@ -36,15 +36,17 @@
     "husky": "^4.2.5",
     "jest": "^26.6.3",
     "recast": "^0.20.4",
-    "rollup": "2.32.1",
+    "rollup": "^2.48.0",
     "rollup-plugin-copy": "^3.3.0",
     "rollup-plugin-eslint": "^7.0.0",
     "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript2": "0.29.0",
-    "rxjs": "^6.6.3",
-    "ts-jest": "^26.4.4",
-    "ts-loader": "8.0.7",
-    "typescript": "4.0.5",
+    "rollup-plugin-typescript2": "^0.30.0",
+    "rxjs": "^7.1.0",
+    "ts-jest": "^26.5.6",
+    "ts-loader": "^9.2.2",
+    "typescript": "^4.2.4",
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.10",
     "ws": "^7.4.5"
   },
   "husky": {
@@ -56,5 +58,8 @@
     "commitizen": {
       "path": "./node_modules/cz-conventional-changelog"
     }
+  },
+  "dependencies": {
+    "tslib": "^2.2.0"
   }
 }
